Encapsulated header artworks filter, implemented to show artworks
